+++ collector/src/main/scala/org/opentorah/collector/CollectionObject.scala
@@ -88,7 +88,7 @@ object CollectionObject {
         yield Ref.toXml(DocumentObject.documentUrl(collection, teiHolder.name), teiHolder.language.get)
       val language: Option[String] = document.tei.languages.map(_.ident).headOption
         .orElse(document.tei.text.lang)
-      Seq(Xml.textNode(language.getOrElse("?"))) ++ translations.flatMap(r => Seq(Xml.textNode(" "), r))
+      Seq(Xml.mkText(language.getOrElse("?"))) ++ translations.flatMap(r => Seq(Xml.mkText(" "), r))
     }),
 
     Table.Column("Документ", "document", { document: Document =>