+++ xml/src/main/scala/org/opentorah/xml/ToXml.scala
@@ -1,20 +1,16 @@
 package org.opentorah.xml
 
-import scala.xml.{Elem, MetaData, Node, Null, UnprefixedAttribute}
+import scala.xml.{Elem, Node}
 
 trait ToXml[A] {
   final def toXml(value: Option[A]): Seq[Elem] = toXml(value.toSeq)
 
   final def toXml(values: Seq[A]): Seq[Elem] = values.map(toXml)
 
-  final def toXml(value: A): Elem = <elem/>.copy(
-    label = elementName(value),
-    attributes = attributes(value).foldRight[MetaData](Null){ case (current, result) => new UnprefixedAttribute(
-      current.attribute.name,
-      current.valueToString.orNull,
-      result
-    )},
-    child = content(value)
+  final def toXml(value: A): Elem = Xml.element(
+    name = elementName(value),
+    attributes = attributes(value),
+    content = content(value)
   )
 
   protected def elementName(value: A): String