+++ CHANGELOG.md
@@ -5,3 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.7.53] - 2020-07-17
+- switched to Logback logging;
+- merged XML attribute functionality;
+- use XML literals instead of strings;
+- cleaned up metadata processing;
+- cleaned up pretty-printing;